cache riverId array
