+++ index.html
@@ -31,10 +31,10 @@
   <script type="module" defer src="/static/js/plots.js"></script>
   <script type="module" defer src="/static/js/reports.js"></script>
   <script type="module" defer src="/static/js/ui.js"></script>
-  <script type="module" defer src="/static/js/data/rfsZarrFetcher.js"></script>
-  <script type="module" defer src="/static/js/data/biasCorrectedApi.js"></script>
-  <script type="module" defer src="/static/js/data/cache.js"></script>
+  <script type="module" defer src="/static/js/data/zarrUtilities.js"></script>
   <script type="module" defer src="/static/js/data/data.js"></script>
+  <script type="module" defer src="/static/js/data/cache.js"></script>
+  <script type="module" defer src="/static/js/data/main.js"></script>
   <script type="module" defer src="/static/js/map.js"></script>
   <script type="module" defer src="/static/js/main.js"></script>
 </head>
