+++ static/js/bookmarks.js
@@ -1,6 +1,6 @@
 import namedDefaultRivers from "../json/namedDefaultRivers.json" with {type: "json"}
 import {RiverId} from "./states/state.js";
-import {validateRiverNumber} from "./data/data.js";
+import {validateRiverNumber} from "./data/main.js";
 
 const key = 'riverBookmarks'
