+++ static/js/main.js
@@ -1,5 +1,5 @@
 import {clearCharts, displayLoadingStatus, displayRiverNumber, divModalCharts, inputForecastDate, lang, riverIdInput, updateDownloadLinks} from "./ui.js";
-import {getAndCacheForecast, getAndCacheRetrospective, getAndCacheReturnPeriods} from "./data/data.js";
+import {getAndCacheForecast, getAndCacheRetrospective, getAndCacheReturnPeriods} from "./data/main.js";
 import {bookmarks} from "./bookmarks.js";
 import {LoadStatus, RiverId, UseBiasCorrected, UseShowExtraRetroGraphs, UseSimpleForecast} from "./states/state.js";
 import {plotAllForecast, plotAllRetro} from "./plots.js";
