+++ static/js/workers/dataFetcher.js
@@ -1,8 +1,8 @@
 // Defines a web worker that listens for a message object with riverId and array of datasets to fetch
-// Uses functions from data.js to fetch and cache the specified datasets for the given riverId
+// Uses functions from main.js to fetch and cache the specified datasets for the given riverId
 // Sends messages to the main thread indicating start and finish progress for progress indicators
 
-import {getAndCacheForecast, getAndCacheRetrospective, getAndCacheReturnPeriods} from '../data/data.js';
+import {getAndCacheForecast, getAndCacheReturnPeriods} from '../data/main.js';
 
 self.onmessage = async function (event) {
   const {riverId, forecastDate} = event.data;
